+++ frontend/src/utils/utils.ts
@@ -5,6 +5,8 @@ import type {
     FontSizeType,
     FontWeightType,
 } from '../types/themeTypes.d.ts';
+import type { MountainCourse, MountainData } from '../types/mountainTypes';
+import type { Option } from '../types/searchBarTypes';
 
 type Background = 'sunny' | 'cloudy' | 'snow' | 'rain';
 
@@ -181,3 +183,64 @@ export const parseFilterFromUrl = (
         return [];
     }
 };
+
+export function refactorMountainDataToOptions(
+    mountainsData: MountainData[],
+): Option[] {
+    return mountainsData.map((mountain) => ({
+        id: mountain.mountainId,
+        name: mountain.mountainName,
+    }));
+}
+
+export function refactorCoursesDataToOptions(
+    coursesData: MountainCourse[],
+): Option[] {
+    return coursesData.map((course) => ({
+        id: course.courseId,
+        name: course.courseName,
+    }));
+}
+
+export function getCurrentTime(): Date {
+    return new Date();
+}
+
+export function formatTimeDifference({
+    pastISO,
+    nowDate = null,
+}: {
+    pastISO: string;
+    nowDate?: Date | null;
+}): string {
+    const now = nowDate ?? getCurrentTime();
+    const past = new Date(pastISO);
+    const diffMs = now.getTime() - past.getTime();
+
+    if (diffMs < 0) return '';
+
+    const diffSec = Math.floor(diffMs / 1000);
+    const diffMin = Math.floor(diffSec / 60);
+    const diffHour = Math.floor(diffMin / 60);
+    const diffDay = Math.floor(diffHour / 24);
+
+    if (diffMin < 60) {
+        return `${diffMin}분 전`;
+    } else if (diffHour < 24) {
+        return `${diffHour}시간 전`;
+    } else {
+        return `${diffDay}일 전`;
+    }
+}
+
+export function filterGatherer({
+    weatherKeywords = [],
+    rainKeywords = [],
+    etceteraKeywords = [],
+}: {
+    weatherKeywords?: string[];
+    rainKeywords?: string[];
+    etceteraKeywords?: string[];
+}) {
+    return [...weatherKeywords, ...rainKeywords, ...etceteraKeywords];
+}
